Moderate Changes: - Update some components - PamphletTitle is now stored in sessionStorage
