+++ src/components/NewSong.tsx
@@ -11,8 +11,6 @@ import {
   Divider,
 } from "@chakra-ui/react";
 import { createSong } from "../util/firestoreFunctions";
-import { useRecoilValue } from "recoil";
-import { sanghefteState } from "../store/store";
 import { useNavigate } from "react-router-dom";
 
 export const NewSong = () => {
@@ -20,16 +18,16 @@ export const NewSong = () => {
   const [title, setTitle] = useState("");
   const [text, setText] = useState("");
   const [artist, setArtist] = useState("");
-  const sanghefte = useRecoilValue(sanghefteState);
   const navigate = useNavigate();
 
+  const pamphletTitle = sessionStorage.getItem("currentPamphlet_title")
   const localStorage_userIdKey = "userID";
 
-  const handleClick = async () => {
+  const handleClick_addSongToPamphlet = async () => {
     const userID = localStorage.getItem(localStorage_userIdKey);
 
-    if (userID !== null) {
-      await createSong(userID, sanghefte, title, text, artist).catch(
+    if (userID !== null && pamphletTitle !== null) {
+      await createSong(userID, pamphletTitle, title, text, artist).catch(
         console.error
       );
       setText("");
@@ -48,11 +46,11 @@ export const NewSong = () => {
   /**
    * Takes the user to the hub if it exists
    */
-  const handleCLick2 = () => {
+  const handleClick_goBackToMyPamphlets = () => {
     const userID = localStorage.getItem(localStorage_userIdKey);
 
     if (userID !== null) {
-      navigate("/");
+      navigate("/pamphletMenu");
     }
   };
 
@@ -67,7 +65,8 @@ export const NewSong = () => {
     >
       <Container>
         <VStack spacing={3}>
-          <Heading>Legg til ny sang</Heading>
+          <Heading>Legg til ny sang i</Heading>
+          <Heading>{pamphletTitle}</Heading>
           <Divider />
           <Heading size="sm"> Tittel </Heading>
           <Input
@@ -90,10 +89,10 @@ export const NewSong = () => {
             value={text}
             onChange={(e) => setText(e.target.value)}
           />
-          <Button onClick={handleClick} isFullWidth colorScheme="teal">
-            Legg til ny sang
+          <Button onClick={handleClick_addSongToPamphlet} isFullWidth colorScheme="teal">
+            Legg til sang
           </Button>
-          <Button isFullWidth onClick={handleCLick2}>
+          <Button isFullWidth onClick={handleClick_goBackToMyPamphlets}>
             Tilbake til Mine Sanghefter
           </Button>
         </VStack>
